voice/requirements.txt
# Voice Segment Generation Pipeline
# Install: pip install -r requirements.txt

# API and HTTP
requests>=2.28.0

# Audio analysis (RMS consistency checking)
numpy>=1.24.0
scipy>=1.10.0

# Pipeline utilities
pyyaml>=6.0
tqdm>=4.60
